Build cleanup: corrected lines which emitted warnings.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@4688 23e1bc29-d954-0410-9766-47de34158ba0

+++ Antlr/base/antlr/String.cpp
@@ -38,7 +38,7 @@ ANTLR_USE_NAMESPACE(std)string operator+( const ANTLR_USE_NAMESPACE(std)string& 
 ANTLR_USE_NAMESPACE(std)string operator+( const ANTLR_USE_NAMESPACE(std)string& lhs, size_t rhs )
 {
 	char tmp[100];
-	sprintf(tmp,"%u",rhs);
+	sprintf(tmp,"%u",static_cast<unsigned int>(rhs));
 	return lhs+tmp;
 }
